add api

+++ Atipico.Application/Interfaces/Services/IEntityService.cs
@@ -2,7 +2,7 @@ namespace Atipico.Application.Interfaces.Services
 {
     public interface IEntityService<TEntity> where TEntity : class
     {
-        Task<TEntity?> GetByIdAsync(int id);
+        Task<TEntity?> GetByIdAsync(long id);
         Task<IEnumerable<TEntity>> GetAllAsync();
         Task<TEntity> AddAsync(TEntity entity);
         Task UpdateAsync(TEntity entity);
