+++ Atipico.Domain/Interfaces/Repositories/IRepository.cs
@@ -7,7 +7,7 @@ namespace Atipico.Domain.Interfaces.Repositories
 {
     public interface IRepository<TEntity> where TEntity : class
     {
-        Task<TEntity?> GetByIdAsync(int id);
+        Task<TEntity?> GetByIdAsync(long id);
         Task<IEnumerable<TEntity>> GetAllAsync();
         Task<IEnumerable<TEntity>> FindAsync(Expression<Func<TEntity, bool>> predicate);
         void Add(TEntity entity);
