+++ Atipico.Infraestructure/Persistence/Repositories/Repository.cs
@@ -33,7 +33,7 @@ namespace Atipico.Infraestructure.Persistence.Repositories
             return await _dbSet.ToListAsync();
         }
 
-        public async Task<TEntity?> GetByIdAsync(int id)
+        public async Task<TEntity?> GetByIdAsync(long id)
         {
             return await _dbSet.FindAsync(id);
         }
